Fix an issue where the parser & serializer badly handled timestamps.
A previous fix changed handling of 'longlong' to use a Buffer instead of a Number, since a Number overflows
at 2^53. It used the same code for Timestamps as well, as they are defined as 64 bit. However, timestamps
(from javascript) are usually passed as Number objects and the serializer/parser expected a buffer, which sent
zeroes.

Timestamp is now parsed & serialized as a Number. 'longlong' is still parsed & serialized as as Buffer, as it
should be expected that if you need this, you know how about javascript's Number type.