Set confirm.select to pre-baked exchanges as well
confirm.select applies to the channel and NOT to the exchange so we have
to set it even for pre-backed exhcanges.

This patch sends the confirmSelect method if the confirm option is set
meaning that the open event (or the callback) will not happen until
confirmSelectOk is received.

This fixes being able to use confirm with default exchanges.