Don't allow publishing on a closed exchange.
Also, ack errors all unacked messages on connection errors in addition
to channel errors.
When the connection goes down, it will only be "detected" during the
next attempt to interact with it. If this happens to be a basicPublish,
then it would fire a callback if not for the fix.